FEniCS/Firedrake backends: Move parameters_key
