+++ tlm_adjoint/firedrake/caches.py
@@ -3,12 +3,12 @@ caching.
 """
 
 from .backend import (
-    Tensor, TrialFunction, backend_assemble, backend_DirichletBC,
+    Parameters, Tensor, TrialFunction, backend_assemble, backend_DirichletBC,
     backend_Function, complex_mode)
 from ..interface import (
     is_var, var_caches, var_id, var_is_cached, var_is_replacement,
     var_replacement, var_space, var_state, weakref_method)
-from .backend_code_generator_interface import linear_solver, parameters_key
+from .backend_code_generator_interface import linear_solver
 
 from ..caches import Cache
 
@@ -19,6 +19,7 @@ from .expr import (
 from .functions import ReplacementFunction
 
 from collections import defaultdict
+from collections.abc import Sequence
 import itertools
 import ufl
 
@@ -343,6 +344,20 @@ def form_dependencies(*forms):
                 deps[dep_id] = dep
 
     return tuple(sorted(deps.values(), key=var_id))
+
+
+def parameters_key(parameters):
+    key = []
+    for name in sorted(parameters.keys()):
+        sub_parameters = parameters[name]
+        if isinstance(sub_parameters, (Parameters, dict)):
+            key.append((name, parameters_key(sub_parameters)))
+        elif isinstance(sub_parameters, Sequence) \
+                and not isinstance(sub_parameters, str):
+            key.append((name, tuple(sub_parameters)))
+        else:
+            key.append((name, sub_parameters))
+    return tuple(key)
 
 
 class AssemblyCache(Cache):
